requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.23.3
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
whitenoise==6.2.0
